+++ src/app/(dashboard)/(admin)/users/schemas/sheets/edit-form.schema.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
-export const UserEditSheetFormSchema = z.object({
+export const UsersEditSheetFormSchema = z.object({
   role: z.enum(['customer', 'editor', 'admin']),
 });
 
-export type UsersEditSheetForm = z.infer<typeof UserEditSheetFormSchema>;
+export type UsersEditSheetForm = z.infer<typeof UsersEditSheetFormSchema>;